requirements.txt
datasets>=2.14.0
transformers>=4.30.0
torch>=2.0.0
sklearn-crfsuite>=0.3.6
pytorch-crf>=0.7.2
seqeval>=1.2.2
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
pyyaml>=6.0
accelerate>=0.26.0
scikit-learn>=1.3.0
umap-learn>=0.5.3
joblib>=1.3.0
